requirements.txt
# Core ML
torch==2.3.1+cpu
torchvision==0.18.1+cpu
numpy==1.26.4
pillow

# API
fastapi
uvicorn
python-multipart

# MLOps
mlflow

# Testing
pytest